fix various compilation warnings

+++ Sources/VLCAppDelegate.h
@@ -31,7 +31,7 @@
 
 - (void)openMediaFromManagedObject:(NSManagedObject *)file;
 - (void)openMovieFromURL:(NSURL *)url;
-- (void)openMediaList:(VLCMediaList*)list atIndex:(NSUInteger)index;
+- (void)openMediaList:(VLCMediaList*)list atIndex:(int)index;
 
 @property (nonatomic, readonly) VLCPlaylistViewController *playlistViewController;
 @property (nonatomic, readonly) VLCDropboxTableViewController *dropboxTableViewController;
